added timing / tree saving

+++ globals.cpp
@@ -1,5 +1,5 @@
 #include "globals.h"
 
 unsigned int sleeptime = 500000;
-float timeLimit = 0;
-float returnTime = 0;
+double timeLimit = 0;
+double returnTime = 0;
